move checkpointer to module
